# Conflicts:
#	Engine/.vs/Engine/v15/.suo
#	Engine/Engine/Code/Framework.cpp
#	Engine/Engine/Code/Framework.h
#	Engine/Engine/Code/Renderer/Renderer.cpp
#	Engine/Engine/Code/Renderer/Renderer.h
#	Engine/Engine/Engine.ilk
#	Engine/Engine/Engine.pdb
#	Engine/Engine/Engine.vcxproj
#	Engine/Engine/Engine.vcxproj.filters
#	Engine/Engine/x64/Debug/Engine.log
#	Engine/Engine/x64/Debug/Engine.tlog/CL.command.1.tlog
#	Engine/Engine/x64/Debug/Engine.tlog/CL.read.1.tlog
#	Engine/Engine/x64/Debug/Engine.tlog/CL.write.1.tlog
#	Engine/Engine/x64/Debug/Engine.tlog/Engine.write.1u.tlog
#	Engine/Engine/x64/Debug/Engine.tlog/link.command.1.tlog
#	Engine/Engine/x64/Debug/Engine.tlog/link.read.1.tlog
#	Engine/Engine/x64/Debug/Engine.tlog/link.write.1.tlog
#	Engine/Engine/x64/Debug/vc141.idb
#	Engine/Engine/x64/Debug/vc141.pdb

+++ Engine/Engine/Code/ComponentSystem/World.cpp
@@ -20,7 +20,7 @@ void World::Update()
 	// update all components one type at the time
 	for (int i = 0; i < NUM_COMPONENT_TYPES; i++) 
 	{
-		unsigned int size = _components[i].size();
+		unsigned int size = (int)_components[i].size();
 
 		for (int y = 0; y < size; y++)
 			if (_components[i][y]->IsActive())
@@ -50,7 +50,7 @@ void World::DeleteEntity(Entity* entity)
 // check if we have any components flagged for removal
 void World::HandleDirtyComponents() 
 {
-	int count = _componentsToRemove.size();
+	int count = (int)_componentsToRemove.size();
 
 	if (count == 0)
 		return;
@@ -66,7 +66,7 @@ void World::HandleDirtyComponents()
 // check if we have any entities flagged for removal
 void World::HandleDirtyEntities()
 {
-	unsigned int count = _entitiesToRemove.size();
+	unsigned int count = (int)_entitiesToRemove.size();
 
 	if (count == 0)
 		return;
@@ -85,7 +85,7 @@ void World::Empty()
 	// delete all components and entitys in the world
 	for (int i = 0; i < NUM_COMPONENT_TYPES; i++) 
 	{
-		unsigned int size = _components[i].size();
+		unsigned int size = (int)_components[i].size();
 
 		for (int y = 0; y < size; y++)
 			delete _components[i][y];
@@ -94,7 +94,7 @@ void World::Empty()
 		_components[i].clear();
 	}
 
-	unsigned int size = _entities.size();
+	unsigned int size = (int)_entities.size();
 
 	for (int i = 0; i < size; i++)
 		delete _entities[i];
